audio analysis logic moved to audioBars component
